move js lib test files to src/js/lib/
