+++ src/js/lib/data_profiles_splitter.test.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai'
-import { DataProfilesSplitter } from '../../src/lib/data_profiles_splitter.js'
+import { DataProfilesSplitter } from './data_profiles_splitter.js'
 
 describe('DataProfilesSplitter', () => {
   const dps = new DataProfilesSplitter();
